requirements.txt
# Web framework and ASGI server
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-dotenv>=1.0.0
Jinja2>=3.0.0
pydantic>=1.10.11
pydantic-settings>=2.0.0
python-multipart>=0.0.5
itsdangerous>=2.1.2

# Database and ORM
sqlalchemy>=2.0.0
alembic>=1.10.0

# HTTP and API clients
requests>=2.28.0
httpx>=0.24.1
cachetools>=5.3.1

# Testing
pytest>=7.0.0

# Modern UI frameworks
nicegui>=1.3.0
reactpy>=1.0.0
reflex>=0.3.0

# Core data science (smaller footprint)
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0

# Web application generation essentials
openai>=0.28.1
PyGithub>=2.6.0

# Optional AI components - Install as needed
# transformers>=4.33.0
# langchain>=0.0.348

# Production server
gunicorn>=20.1.0

# Data visualization
plotly>=5.14.0
matplotlib>=3.7.0

# Async capabilities
asyncio>=3.4.3
aiohttp>=3.8.4